📃 chore: サーバーサイド用のaxiosは一度しか使っていなかったので、統合
