+++ src/hooks/useDeleteUser.ts
@@ -1,7 +1,7 @@
 import { useRouter } from 'next/navigation';
 import toast from 'react-hot-toast';
 import { signOut } from 'next-auth/react';
-import { clientAxiosDelete } from '@/lib/clientAxios';
+import { axiosDelete } from '@/lib/axios';
 import { useSession } from 'next-auth/react';
 
 const useDeleteUser = () => {
@@ -11,7 +11,7 @@ const useDeleteUser = () => {
   const handleDeleteUser = async () => {
     const token = session?.user?.accessToken;
     try {
-      const res = await clientAxiosDelete(`/users/${session?.user?.id}`, token);
+      const res = await axiosDelete(`/users/${session?.user?.id}`, token);
       if (res.status === 204) {
         signOut();
         router.push('/thanks');
