+++ src/lib/axios.ts
@@ -1,17 +1,43 @@
 import axios from 'axios';
-import { auth } from '@/auth';
 
 const BASE_URL = process.env.NEXT_PUBLIC_RAILS_API_URL;
 
-async function getHeaders() {
-  const session = await auth();
-  const token = session?.user?.accessToken ?? '';
-  return {
+export async function axiosGet(url: string, token: string | undefined) {
+  const headers = {
     Authorization: `Bearer ${token}`,
   };
+
+  return await axios.get(`${BASE_URL}${url}`, { headers });
 }
 
-export async function axiosGet(url: string) {
-  const headers = await getHeaders();
-  return await axios.get(`${BASE_URL}${url}`, { headers });
+export async function axiosPost<B>(
+  url: string,
+  token: string | undefined,
+  body: B,
+) {
+  const headers = {
+    Authorization: `Bearer ${token}`,
+  };
+
+  return await axios.post(`${BASE_URL}${url}`, body, { headers });
+}
+
+export async function axiosPatch<B>(
+  url: string,
+  token: string | undefined,
+  body: B,
+) {
+  const headers = {
+    Authorization: `Bearer ${token}`,
+  };
+
+  return await axios.patch(`${BASE_URL}${url}`, body, { headers });
+}
+
+export async function axiosDelete(url: string, token: string | undefined) {
+  const headers = {
+    Authorization: `Bearer ${token}`,
+  };
+
+  return await axios.delete(`${BASE_URL}${url}`, { headers });
 }
