+++ src/utils/getBooks.ts
@@ -1,6 +1,9 @@
 import { axiosGet } from '@/lib/axios';
+import { auth } from '@/auth';
 
 export default async function getBooks() {
-  const res = await axiosGet('/user_books');
+  const session = await auth();
+  const token = session?.user?.accessToken;
+  const res = await axiosGet('/user_books', token);
   return res.data;
 }
